remove the ClearButton combine with BpkInput
