+++ .eslintrc
@@ -139,7 +139,6 @@
             "allowedFor": [
               "ArrowLeftIcon", // allowed for icon to continue supporting RTL with css
               "ArrowRightIcon", // allowed for icon to continue supporting RTL with css
-              "BpkClearButton", //  allowed for a prop that is passed down in the BpkInput to a generated element
               "BpkSelectableChip", // allowed for prop that is part of the classnames applied to the BpkSelectableChip
               "CornerRadius", // allowed for a prop that is passed down in the BpkContentBubble as partof BpkFlare to a generated element and necessary to continue supporting svgClassName
               "DateContainer", // allowed for a prop that is passed down in the BpkCalendarWeek to the DateContainer
